feat: superjson checksum
* refactored the whole superjson file so now we can read it
