+++ src/client/profile-provider.test.ts
@@ -15,7 +15,8 @@ import {
   ProviderJson,
   SecurityType,
 } from '../internal/providerjson';
-import { SecurityValues, SuperJson } from '../internal/superjson';
+import { SuperJson } from '../internal/superjson';
+import * as SuperJsonMutate from '../internal/superjson/mutate';
 import { err, ok } from '../lib';
 import { ProfileConfiguration } from './profile';
 import { BoundProfileProvider, ProfileProvider } from './profile-provider';
@@ -40,8 +41,6 @@ jest.mock('fs', () => ({
 }));
 
 //Mock super json
-jest.mock('../internal/superjson');
-
 const mockResolvePath = jest.fn();
 jest.mock('../internal/superjson', () => ({
   ...jest.requireActual<Record<string, unknown>>('../internal/superjson'),
@@ -681,31 +680,28 @@ describe('profile provider', () => {
       });
 
       it('returns new BoundProfileProvider with merged security', async () => {
-        // eslint-disable-next-line @typescript-eslint/unbound-method
-        const orginalMerge = SuperJson.mergeSecurity;
-
-        SuperJson.mergeSecurity = (
-          _left: SecurityValues[],
-          _right: SecurityValues[]
-        ) => [
-          {
-            username: 'test-username',
-            id: 'basic',
-            password: 'test-password',
-          },
-          {
-            id: 'api',
-            apikey: 'test-api-key',
-          },
-          {
-            id: 'bearer',
-            token: 'test-token',
-          },
-          {
-            id: 'digest',
-            digest: 'test-digest-token',
-          },
-        ];
+        const mergeSecuritySpy = jest.spyOn(SuperJsonMutate, 'mergeSecurity');
+        mergeSecuritySpy.mockReturnValue(
+          [
+            {
+              username: 'test-username',
+              id: 'basic',
+              password: 'test-password',
+            },
+            {
+              id: 'api',
+              apikey: 'test-api-key',
+            },
+            {
+              id: 'bearer',
+              token: 'test-token',
+            },
+            {
+              id: 'digest',
+              digest: 'test-digest-token',
+            },
+          ]
+        );
         mocked(fetchBind).mockResolvedValue(mockFetchResponse);
 
         //normalized is getter on SuperJson - unable to mock or spy on
@@ -748,7 +744,7 @@ describe('profile provider', () => {
             ],
           }).toString()
         );
-        SuperJson.mergeSecurity = orginalMerge;
+        mergeSecuritySpy.mockRestore();
       });
 
       it('throws error when could not find scheme', async () => {
