+++ src/client/provider.ts
@@ -1,4 +1,5 @@
-import { SecurityValues, SuperJson } from '../internal';
+import { SecurityValues } from '../internal';
+import { mergeSecurity } from '../internal/superjson/mutate';
 import { SuperfaceClientBase } from './client';
 
 export class ProviderConfiguration {
@@ -26,7 +27,7 @@ export class Provider {
   }): Promise<Provider> {
     const newConfiguration = new ProviderConfiguration(
       this.configuration.name,
-      SuperJson.mergeSecurity(
+      mergeSecurity(
         this.configuration.security,
         configuration.security ?? []
       ),
