+++ src/lib/object.ts
@@ -5,9 +5,16 @@ export function clone<T>(value: T): T {
   return JSON.parse(JSON.stringify(value)) as T;
 }
 
+export function isRecord(input: unknown): input is Record<string, unknown> {
+  if (typeof input !== 'object' || input === null) {
+    return false;
+  }
+
+  return true;
+}
+
 /**
- * Recursively descends the record and returns a list of enumerable all keys
- *
+ * Recursively descends the record and returns a list of enumerable keys
  */
 export function recursiveKeyList(
   record: Record<string, unknown>,
@@ -34,4 +41,35 @@ export function recursiveKeyList(
   }
 
   return keys;
+}
+
+/**
+ * Recursively index into a record.
+ * 
+ * Throws if a child cannot be indexed into.
+ */
+export function indexRecord<T extends unknown | Record<string, T>>(
+  input: Record<string, T>,
+  key: string[]
+): T | undefined {
+  // check for input being undefined is for sanity only
+  if (key.length === 0 || input === null || input === undefined) {
+    return undefined;
+  }
+
+  if (key.length === 1) {
+    return input[key[0]];
+  }
+
+  const currentKey = key.shift();
+  if (currentKey === undefined) {
+    throw new Error('unreachable');
+  }
+
+  const next = input[currentKey];
+  if (!isRecord(next)) {
+    throw new Error('Cannot index into non-object');
+  }
+
+  return indexRecord(next as Record<string, T>, key)
 }
